requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
cssselect==1.2.0
cssutils==2.9.0
Django==4.2.7
django-anymail==10.2
django-constance==3.1.0
django-enumfield==3.1
django-enumfields==2.1.1
django-mock-queries==2.2.0
django-picklefield==3.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==20.1.0
flower==2.0.1
freezegun==1.2.2
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
kombu==5.3.4
lxml==4.9.3
model-bakery==1.17.0
packaging==23.2
pluggy==1.3.0
premailer==3.10.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
sendgrid==6.11.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tomli==2.0.1
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
